+++ hw/riscv/ibex_common.c
@@ -584,11 +584,11 @@ uint32_t ibex_load_kernel(AddressSpace *as)
     return (uint32_t)kernel_entry;
 }
 
-uint64_t ibex_get_current_pc(void)
+uint32_t ibex_get_current_pc(void)
 {
     CPUState *cs = current_cpu;
 
-    return cs && cs->cc->get_pc ? cs->cc->get_pc(cs) : 0u;
+    return cs && cs->cc->get_pc ? (uint32_t)cs->cc->get_pc(cs) : 0u;
 }
 
 /* x0 is replaced with PC */
